code/customers/views.py

from .models import Customer
from .serializers import CustomerSerializer
from rest_framework import viewsets

class CustomerViewSet(viewsets.ModelViewSet):
    queryset = Customer.objects.all()
    serializer_class = CustomerSerializer