code/requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.19
django-cors-headers==3.7.0
django-filter==2.4.0
django-nvd3==0.9.7
django-oauth-toolkit==1.4.1
django-rest-framework-proxy==1.4.0
django-totalsum-admin==1.0.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
factory-boy==3.1.0
Faker==6.6.0
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.3
Markdown==2.6.11
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.1.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
Unidecode==1.2.0
urllib3==1.26.3